Update cmake settings and reduce warnings

+++ src/Akuma/ProceduralGen/BSP/Node.cpp
@@ -16,8 +16,8 @@ BSP::Node::~Node() {
 glm::vec2 BSP::Node::getCentrePoint() {
     glm::vec2 centre;
     centre = (gridTopRight - gridBottomLeft);
-    centre.x *= 0.5;
-    centre.y *= 0.5;
+    centre.x *= static_cast<float>(0.5);
+    centre.y *= static_cast<float>(0.5);
     centre += gridBottomLeft;
     return centre;
 }
